regroup queue methods
