--- kuzzle/flush_queue.go
@@ -1,6 +1,0 @@
-package kuzzle
-
-// FlushQueue empties the offline queue without replaying it.
-func (k *Kuzzle) FlushQueue() {
-	k.socket.ClearQueue()
-}
