+++ kuzzle/queue_test.go
