--- kuzzle/start_queuing.go
@@ -1,5 +1,0 @@
-package kuzzle
-
-func (k *Kuzzle) StartQueuing() {
-	k.socket.StartQueuing()
-}
