--- kuzzle/stop_queuing.go
@@ -1,5 +1,0 @@
-package kuzzle
-
-func (k *Kuzzle) StopQueuing() {
-	k.socket.StopQueuing()
-}
